Rename adf_dev_dump to adf_dev_driver_dump, lclevy#73
