+++ src/adf_dev_hd.c
@@ -28,7 +28,7 @@
 
 #include "adf_dev_hd.h"
 
-#include "adf_dev_dump.h"
+#include "adf_dev_driver_dump.h"
 #include "adf_env.h"
 #include "adf_raw.h"
 #include "adf_util.h"
