+++ src/adf_env.c
@@ -29,7 +29,7 @@
 
 #include "adf_blk.h"
 #include "adf_dev_drivers.h"
-#include "adf_dev_dump.h"
+#include "adf_dev_driver_dump.h"
 #include "adf_dev_ramdisk.h"
 #include "adf_version.h"
 #include "defendian.h"
